ACC-280: Rename tax classes
