--- src/Countries/US/California/SanFranciscoPayrollEmployer/V20190101/SanFranciscoPayrollEmployer.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Appleton\Taxes\Countries\US\California\SanFranciscoPayrollEmployer\V20190101;
-
-use Appleton\Taxes\Countries\US\California\SanFranciscoPayrollEmployer\SanFranciscoPayrollEmployer as BaseSanFranciscoPayrollEmployer;
-
-class SanFranciscoPayrollEmployer extends BaseSanFranciscoPayrollEmployer
-{
-    private const START_AMOUNT = 30000000;
-    private const TAX_AMOUNT = 0.0038;
-
-    public function getStartAmount(): int
-    {
-        return self::START_AMOUNT;
-    }
-
-    public function getTaxAmount(): float
-    {
-        return self::TAX_AMOUNT;
-    }
-}
